add graph solution plot

+++ VRP/src/graphic/Node.java
@@ -5,11 +5,11 @@ public class Node {
 	
 	public Object racine;
 	public int num;
-	public int pos_X;
-	public int pos_Y;
+	public double pos_X;
+	public double pos_Y;
 	
 	
-	public Node(int x, int y, int num) {
+	public Node(double x, double y, int num) {
 		racine = null;
 		this.num = num;
 		this.pos_X = x;
